fixed minor bugs in screening controller and edit view

+++ CinemaApplication/Controllers/ScreeningController.cs
@@ -21,6 +21,13 @@ namespace CinemaApplication.Controllers
         }
 
         [HttpGet]
+        public IActionResult IndexByDate(DateTime minTime, DateTime maxTime)
+        {
+            IEnumerable<Screening> screenings = _db.Screenings.Where(s => s.Time >= minTime && s.Time <= maxTime);
+            return View(screenings);
+        }
+
+        [HttpGet]
         public IActionResult Index(String MovieName)
         {
             TempData["MovieName"] = MovieName;
@@ -29,7 +36,7 @@ namespace CinemaApplication.Controllers
                 .Where(s => s.Movie.MovieName == MovieName && s.Time > DateTime.Now)
                 .Include(s => s.Cinema)
                 .Include(s => s.Bookings)
-                .OrderBy(s=>s.Time);
+                .OrderBy(s => s.Time);
 
             // booking available seats are assigned cinema seats minus the amount of bookings* seats per booking
             foreach (var screening in movieScreenings)
@@ -39,23 +46,34 @@ namespace CinemaApplication.Controllers
             return View(movieScreenings);
         }
         [HttpPost]
-        public IActionResult Create(String MovieName,String CinemaName,DateTime Time)
+        public IActionResult Create(String MovieName, String CinemaName, DateTime Time)
         {
             if (!HttpContext.Session.GetString("role").Equals("CONTENT-ADMIN"))
             {
                 return RedirectToAction("Index", "Movie");
             }
             TempData["role"] = HttpContext.Session.GetString("role");
-            Cinema cinema = _db.Cinemas.FirstOrDefault(c=>c.Name == CinemaName);
-            Movie movie = _db.Movies.FirstOrDefault(m=>m.MovieName==MovieName);
-            ContentAdmin contentAdmin = _db.Set<ContentAdmin>().FirstOrDefault(u => u.Username.Equals(HttpContext.Session.GetString("username")));
+            Cinema cinema = _db.Cinemas.Find(CinemaName);
+            if (cinema == null)
+            {
+                TempData["error"] = "Cinema does not exist";
+                return RedirectToAction("Index", new { MovieName = MovieName });
+            }
+            Movie movie = _db.Movies.Find(MovieName);
+            if (movie == null)
+            {
+                TempData["error"] = "Movie does not exist";
+                return RedirectToAction("Index", "Movie");
+            }
+            ContentAdmin contentAdmin = _db.Set<ContentAdmin>().Find(HttpContext.Session.GetString("username"));
+            
             Screening screening = new Screening();
             if (!_db.Screenings
-                .Include(s=>s.Movie)
-                .Include(s=>s.Cinema)
+                .Include(s => s.Movie)
+                .Include(s => s.Cinema)
                 .Where(
-                    s=>(DateTime.Compare(s.Time.AddMinutes(movie.Length),Time)>=0 
-                    && DateTime.Compare(Time.AddMinutes(movie.Length), s.Time) >= 0) 
+                    s => (DateTime.Compare(s.Time.AddMinutes(s.Movie.Length), Time) >= 0
+                    || DateTime.Compare(Time.AddMinutes(movie.Length), s.Time) >= 0) 
                     && s.Cinema.Name.Equals(CinemaName))
                 .IsNullOrEmpty())   //checks whether this screening's time and duration overlap with existing one 
             {
@@ -105,6 +123,11 @@ namespace CinemaApplication.Controllers
             }
             TempData["role"] = HttpContext.Session.GetString("role");
             Screening screening = _db.Screenings.Include(s => s.Bookings).Include(s => s.Movie).FirstOrDefault(s => s.Id == Id);
+            if (screening == null)
+            {
+                TempData["error"] = "Screening does not exist , try again";
+                return RedirectToAction("Index", "Movie");
+            }
             _db.Bookings.RemoveRange(screening.Bookings);
             String movieName = screening.Movie.MovieName;
             _db.Screenings.Remove(screening);
